added working source
Working source for sending commands to the PI-bot, PID control not ok!
